Move RunWidget to subfolder of Common
